Resolve warnings and lints
